+++ tests/ThemeTest.php
@@ -492,7 +492,7 @@ END;
         });
     }
 
-    public function test_register_view_to_render_before_and_after()
+    public function test_theme_register_views_event_to_insert_views_before_and_after()
     {
         $this->usingThemeFolder(function (string $folder) {
             $before = 'this-is-my-before-header-string';
@@ -502,10 +502,14 @@ END;
 
             $functionsContent = <<<'CONTENT'
 <?php use BookStack\Facades\Theme;
-Theme::registerViewToRenderBefore('layouts.parts.header', 'before', 4);
-Theme::registerViewToRenderAfter('layouts.parts.header', 'after-a', 4);
-Theme::registerViewToRenderAfter('layouts.parts.header', 'after-b', 1);
-Theme::registerViewToRenderAfter('layouts.parts.header', 'after-c', 12);
+use BookStack\Theming\ThemeEvents;
+use BookStack\Theming\ThemeViews;
+Theme::listen(ThemeEvents::THEME_REGISTER_VIEWS, function (ThemeViews $themeViews) {
+    $themeViews->renderBefore('layouts.parts.header', 'before', 4);
+    $themeViews->renderAfter('layouts.parts.header', 'after-a', 4);
+    $themeViews->renderAfter('layouts.parts.header', 'after-b', 1);
+    $themeViews->renderAfter('layouts.parts.header', 'after-c', 12);
+});
 CONTENT;
 
             $viewDir = theme_path();
@@ -516,12 +520,15 @@ CONTENT;
             file_put_contents($viewDir . '/after-c.blade.php', $afterC);
 
             $this->refreshApplication();
+            $this->artisan('view:clear');
 
             $resp = $this->get('/login');
             $resp->assertSee($before);
             // Ensure ordering of the multiple after views
             $resp->assertSee($afterB . "\n" . $afterA . "\nthis-is-my-after-header-string-52");
         });
+
+        $this->artisan('view:clear');
     }
 
     protected function usingThemeFolder(callable $callback)
